refactor(examples-next-js): improve structure
