+++ examples/with-next-js/src/app/layout.tsx
@@ -3,13 +3,13 @@ import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import { PropsWithChildren } from 'react';
 
-import './styles/global.css';
+import '../styles/global.css';
 
 const inter = Inter({ subsets: ['latin'] });
 
 export const metadata: Metadata = {
   title: 'Create Next App',
-  description: 'Generated by create next app',
+  description: 'Generated by create-next-app',
 };
 
 function RootLayout({ children }: PropsWithChildren) {
