+++ examples/with-next-js/src/services/github.ts
