+++ examples/with-next-js/src/styles/global.css
